Projeto Final

+++ Projeto_Final/TropPizza/server/TropPizza.Infra.Data/Repositories/OrderRepository.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using TropPizza.Domain.Exceptions;
 using TropPizza.Domain.Features.Orders;
+using TropPizza.Domain.Features.Orders.Enums;
 using TropPizza.Domain.Features.Products;
 using TropPizza.Infra.Data.DAO;
 
@@ -10,28 +11,29 @@ namespace TropPizza.Infra.Data.Repositories
     public class OrderRepository : IOrderRepository
     {
         private CustomerRepository _customerRepository = new CustomerRepository();
+        private ProductRepository _productRepository = new ProductRepository();
         private OrderProductsRepository _orderProductsRepository = new OrderProductsRepository();
         private OrderDAO _orderDAO = new OrderDAO();
 
         public void Create(Order order)
         {
-            // Order searchedOrder = _orderDAO.ReadById(order.Id); // get order ID!
-            // if (searchedOrder != null)
-            // {
-            //     throw new AlreadyExists();
-            // }
-
             if (order.Validate())
             {
-                if (!String.IsNullOrEmpty(order.CustomerCpf))
+                if (order.OrderCustomer != null)
                 {
-                    _customerRepository.ApplyFidelityPoints(order.CustomerCpf, order.TotalPrice);
+                    _customerRepository.ApplyFidelityPoints(order.OrderCustomer.Id, order.TotalPrice);
                 }
 
                 _orderDAO.Create(order);
                 Int64 lastKey = _orderDAO.GetLastKey();
                 _orderProductsRepository.Create(order, lastKey);
-                _orderProductsRepository.RemoveFromInventory(order.Products); // de quem é essa responsabilidade?
+
+                foreach (Product product in order.Products)
+                {
+                    Product stockProduct = _productRepository.ReadById(product.Id);
+                    stockProduct.RemoveFromInventory(product.Quantity);
+                    _productRepository.Update(stockProduct);
+                }
             }
         }
 
@@ -71,14 +73,30 @@ namespace TropPizza.Infra.Data.Repositories
                 throw new NotFound();
             }
 
-            // if (order.CustomerCpf)
-            // {
-            //     _customerRepository.RemoveFidelityPoints(order.CustomerCpf, order.TotalPrice);
-            // }
+            searchedOrder.Products = _orderProductsRepository.ReadById(id);
 
-            // _orderProductsRepository.AddToInventory(order.Products); // de quem é essa 
-            _orderProductsRepository.Delete(id);
-            _orderDAO.Delete(id);
+            if (searchedOrder.CanBeDeleted())
+            {
+                if (searchedOrder.OrderCustomer != null)
+                {
+                    _customerRepository.ApplyFidelityPoints(searchedOrder.OrderCustomer.Id, -searchedOrder.TotalPrice);
+                }
+
+                if (searchedOrder.Status == OrderStatus.Pending)
+                {
+                    foreach (Product product in searchedOrder.Products)
+                    {
+                        Product stockProduct = _productRepository.ReadById(product.Id);
+                        stockProduct.AddToInventory(product.Quantity);
+                        _productRepository.Update(stockProduct);
+                    }
+                }
+
+                _orderProductsRepository.Delete(id);
+                _orderDAO.Delete(id);
+            } else {
+                
+            }
         }
     }
 }
